+++ tests/test_models.py
@@ -16,6 +16,8 @@ PHASE1_TABLES = {
 EVENT_INTELLIGENCE_FOUNDATION_TABLES = {
     "event_candidates",
     "event_candidate_evidence",
+    "impact_analyses",
+    "event_fact_snapshots",
 }
 
 FUTURE_TABLE_NAMES = {